Added a Page object (see Page.h and Page.cpp) which will store a finite number of (string) lineItems as well as provide for a header and footer.
Also added the "skeleton" for a new Database.h function which will query the AccountTransaction table for a given accountNumber, and build each Page object one at a time.  Each page will be added to a vector of Pages and once the vector is complete, the function will return a pointer to the vector so the UI can display the resulting collection.