Wrote a function which confirms how the KompexWrapper deals with retrieving multiple rows of data.
See the function printTxHxToConsole() in the Database.h file and run just the two lines in Source.cpp within the "Database Tests Related to Transaction History" region (comment out the rest of the main method) to see this in action.

Using these findings, I'm going to move forward with defining the logic for the getTransactionHistory() function.