Added a new Main Menu option for users to view Transfer History.
Transfers need a separate "history option" for now because of the differences in the Transaction and Transfer History tables.  Transfers have the added problem of keeping track of source and destination accounts.

The transfer history report definitely needs cleaning up as it is currently just re-using the transaction history format.