Changed the Account table to include a column for "status".
Added a function which allows the user to change the status of their account from ACTIVE to INACTIVE.

If the user changes the account status from ACTIVE to INACTIVE, they are immediately logged out and cannot log in at all anymore while the status remains active.

Currently there is no way to undo this action once it has been done. But for this application there doesn't need to be a way to undo it so it doesn't really matter too much.

I'm pretty satisfied wth the project as it stands currently.  Obviously there are a ton of ways this can be improved, but for now it looks pretty okay.

My next task is to thoroughly comment everything that isn't explained already and make edits where necessary.